Fix gamemode Fix break time (bug with mining fatigue)

+++ gomint-api/src/main/java/io/gomint/inventory/item/ItemReduceBreaktime.java
@@ -13,6 +13,6 @@ public interface ItemReduceBreaktime extends ItemStack {
      *
      * @return divisor for the formula
      */
-    double getDivisor();
+    float getDivisor();
 
 }
