+++ gomint-server/src/main/java/io/gomint/server/inventory/item/ItemReduceTierDiamond.java
@@ -25,7 +25,7 @@ public abstract class ItemReduceTierDiamond extends ItemStack implements ItemRed
     }
 
     @Override
-    public double getDivisor() {
+    public float getDivisor() {
         return 8;
     }
