+++ gomint-server/src/main/java/io/gomint/server/inventory/item/ItemReduceTierSpecial.java
@@ -32,7 +32,7 @@ public abstract class ItemReduceTierSpecial extends ItemStack implements ItemRed
     }
 
     @Override
-    public double getDivisor() {
+    public float getDivisor() {
         return 15;
     }
